+++ common/src/main/java/com/dudko/blazinghot/gui/BlazingGuiTextures.java
@@ -8,7 +8,8 @@ import net.minecraft.resources.ResourceLocation;
 
 public enum BlazingGuiTextures {
 
-	JEI_SHORT_ARROW_LEFT("jei/widgets", 15, 9);
+	JEI_SHORT_ARROW_LEFT("jei/widgets", 0, 0, 15, 9),
+	JEI_PLUS("jei/widgets", 16, 0, 9, 9);
 
 	public final ResourceLocation location;
 	public final int width, height;